Stabilize switching between models
